- Move the #178 test case to the 'HotColumn' section - Add a test for mixing the key-enabled and default editors between columns.
